requirements.txt
addict==2.4.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
appnope==0.1.4
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
async-lru==2.0.5
attr==0.3.1
attrs==25.3.0
azure-core==1.35.1
azure-storage-blob==12.26.0
babel==2.17.0
beautifulsoup4==4.13.5
black==25.1.0
bleach==5.0.1
boto==2.49.0
boto3==1.40.30
botocore==1.40.30
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
datamodel-code-generator==0.26.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.12
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email-validator==2.3.0
executing==2.2.1
expiringdict==1.2.2
Faker==37.6.0
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.59.2
fqdn==1.5.1
fsspec==2025.9.0
genson==1.3.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
groundingdino-py==0.4.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.4.0
importlib_metadata==8.7.0
inflect==5.6.2
inflection==0.5.1
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsf==0.11.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
label-studio==1.20.0
label-studio-sdk==1.0.18
lark==1.2.2
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==6.0.1
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
openai==1.107.2
opencv-python==4.11.0.86
opentelemetry-api==1.37.0
ordered-set==4.0.2
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.1
prometheus_client==0.22.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycocotools==2.0.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.4
pyRFC3339==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.2
pyzmq==27.1.0
redis==5.2.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
requests-mock==1.12.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
rq==1.16.2
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semver==3.0.4
Send2Trash==1.8.3
sentry-sdk==2.37.1
setuptools==80.9.0
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
supervision==0.6.0
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
timm==1.0.19
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
ultralytics==8.3.199
ultralytics-thop==2.0.17
uri-template==1.3.0
uritemplate==4.2.0
urllib3==1.26.20
user-agents==2.2.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.40.0
widgetsnbextension==4.0.14
wrapt==1.17.3
xmljson==0.2.1
yapf==0.43.0
zipp==3.23.0
